+++ src/proto_tcl.h
@@ -8,8 +8,8 @@
 
 struct tcl_handle {
 	/* derived from ats */
-	unsigned int fsc;	/* max frame size accepted by card */
-	unsigned int fsd;	/* max frame size accepted by reader */
+	size_t fsc;	/* max frame size accepted by card */
+	size_t fsd;	/* max frame size accepted by reader */
 	unsigned int fwt;	/* frame waiting time (in usec)*/
 	unsigned char ta;	/* divisor information */
 	unsigned char sfgt;	/* start-up frame guard time (in usec) */
